move sourcemaps out of deploy-helpers
