+++ packages/deploy-helpers/src/deploy/deploy.ts
@@ -50,7 +50,6 @@ import {
 	addRequiredSecretsInheritBindings,
 	handleMissingSecretsError,
 } from "./helpers/secrets-validation";
-import { loadSourceMaps } from "./helpers/source-maps";
 import {
 	getSourceMappedString,
 	maybeRetrieveFileSourceMap,
@@ -168,7 +167,6 @@ export default async function deploy(
 		compatibilityDate,
 		compatibilityFlags,
 		keepVars,
-		uploadSourceMaps,
 		accountId,
 	} = props;
 
@@ -404,7 +402,7 @@ See https://developers.cloudflare.com/workers/platform/compatibility-dates for m
 		resolvedEntryPointPath,
 		bundleType,
 		content,
-		bundle,
+		sourceMaps,
 	} = buildResult;
 	// durable object migrations
 	const migrations = !isDryRun
@@ -504,9 +502,7 @@ See https://developers.cloudflare.com/workers/platform/compatibility-dates for m
 		migrations,
 		modules,
 		containers: config.containers,
-		sourceMaps: uploadSourceMaps
-			? loadSourceMaps(main, modules, bundle)
-			: undefined,
+		sourceMaps,
 		compatibility_date: compatibilityDate,
 		compatibility_flags: compatibilityFlags,
 		keepVars,
